:pencil2: board week#3 march - refactoring firebase connection, cleaning, common things to common place
